requirements.txt
pandas
pandas-market-calendars
numpy
pyarrow
fastparquet
numba
ccxt
pydantic<2
click
tqdm
sqlalchemy
python-dotenv
pyyaml
requests
fastapi
prometheus-client
uvicorn
httpx
pytest
